requirements.txt
Django==1.6.1
Pillow==2.3.0
bleach==1.4
django-allauth==0.15.0
django-autoslug==1.7.2
django-avatar==2.0
django-bootstrap-form==3.1
django-classy-tags==0.4
django-compressor==1.3
django-mptt==0.6.0
django-permissionsx==1.3.0
django-sekizai==0.7
django-tastypie==0.11.0
django-tinymce==1.5.2
html5lib==0.999
oauthlib==0.6.1
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
six==1.5.2
wsgiref==0.1.2
